Add Test for verbose logging on github.Client
